modified:   cmake-build-debug/.ninja_deps 	modified:   cmake-build-debug/.ninja_log 	modified:   cmake-build-debug/CMakeFiles/output.dir/src/Polygon.cpp.o 	modified:   cmake-build-debug/CMakeFiles/output.dir/src/PolygonClipping.cpp.o 	modified:   cmake-build-debug/CMakeFiles/output.dir/src/main.cpp.o 	modified:   cmake-build-debug/Testing/Temporary/LastTest.log 	modified:   cmake-build-debug/output 	modified:   src/Polygon.cpp 	modified:   src/Polygon.h 	modified:   src/PolygonClipping.cpp 	modified:   src/PolygonClipping.h 	modified:   src/PolygonUtils.h 	modified:   src/ProcessPolygon.h 	modified:   src/main.cpp

+++ src/Polygon.cpp
@@ -76,7 +76,7 @@ void PolyClip::Polygon::Insert(Vertex *new_v, Vertex *pos) {
     vertexNum_++;
 }
 
-void PolyClip::Polygon::Insert(float x, float y, Vertex *pos) {
+void PolyClip::Polygon::Insert(double x, double y, Vertex *pos) {
     auto * new_v = new Vertex(x, y);
     auto next_v = pos->next_;
